MDCv1-package {MDCv1}	R Documentation

Modified discrete choice model

Description

Fits the Modified Discrete Choice (MDC) model to estimate resource selection when GPS fix success < 1.

Details

When GPS fix success is <1 and related to habitat charachteristics (e.g., canopy cover), typical estimating RSFs can result of biased estimates of resource selection. The modified discrete choice (MDC) model accounts for missing locations.

When missing locations are not an issue, the standard discrete choice (DC) is used to model resource selection.

One option included in both modeling approaches is to consider distance from previous location when estimating resource selection. By including 'distance' from previous location in the model, we are considering that the animal does not move too far from the previous location when making its next selection. Considering animal movement when estimating resource selection is the basis for step-selection functions.

Author(s)

Ryan Nielson <ryan@eagleenvironmental.net> , John Lombardi, and Trent McDonald

References

Nielson, R. M., B. F. J. Manly, L. L. McDonald, H. Sawyer, and T. L. McDonald. 2009. Estimating habitat selection when GPS fix success is less than 1. Ecology 90:2956-2962.

See Also

mdc

[Package MDCv1 version 1.0 Index]