locations {MDCv1}	R Documentation

Animal locations (cell_IDs)

Usage

data("locations")

Format

A data frame with 444 observations on the following 2 variables.

unit.id

a numeric vector of cell_IDs

fix

a numeric vector of fix attempt ID

Details

cell_IDs should represent every fix attempt, so NA's exist when the animal wasn't located. Cell_IDs should be in sequential order. The 'fix' column is not necessary for analysis, but demonstrates that unsuccessful fix attempts are represented in the data.

Source

Nielson, R. M., B. F. J. Manly, L. L. McDonald, H. Sawyer, and T. L. McDonald. 2009. Estimating habitat selection when GPS fix success is less than 1. Ecology 90:2956-2962.

Examples


str(locations)

[Package MDCv1 version 1.0 Index]