mdc {MDCv1}	R Documentation

Fits the modified discrete choice model

Description

Just what the title says

Usage


mdc(habitatDF, CellID, rsfCovars = NULL, probDetCovars = NULL, distColumns = NULL, maxLagArg = 3, SEs = NULL)

Arguments

habitatDF
data.frame of habitat units (cells) across the landscape, with X and Y coordinates, CellID, and habitat covariates.

CellID
Vector of CellIDs for animal relocations, in sequential order, with NA values when the animal was missing.

rsfCovars
Character string of covariate names to be used in the resource selection function piece of the MDC. Include 'distance' if distance from previous location should be included in the RSF.

probDetCovars
Character string of covariate names to be used in the probability of detection piece of the MDC. Set to NULL for fitting a standard discrete choice model.

distColumns
Names of X and Y coordinates in habitatDF, or NULL.

maxLagArg
Do Not Disturb

SEs
Do you want SEs to be estimated using the Hessian? Must be TRUE or FALSE. Saying FALSE saves time (about 30 percent reduction), but you don't get SEs! This is really intended for speedy testing.

Value

Results of model estimation.

See Also

MDCv1 habitat locations

Examples


habitat <- MDCv1::habitat
locations <- MDCv1::locations

# Fix success rate
mean(!is.na(locations$unit.id))

# Standard discrete choice model
mdc(habitatDF = habitat, CellID = locations$unit.id, 
    rsfCovars = c("distance", "prcnt.sage", "elevation"), 
    probDetCovars = NULL, distColumns = c("utmX","utmY"), 
    maxLagArg = 3, SEs = TRUE)

# Modified discrete choice model
mdc(habitatDF = habitat, CellID = locations$unit.id, 
    rsfCovars = c("distance", "prcnt.sage", "elevation"), 
    probDetCovars = "prcnt.sage", distColumns = c("utmX","utmY"), 
    maxLagArg = 3, SEs = TRUE)

[Package MDCv1 version 1.0 Index]